+++ src/main/java/com/team5/projrental/common/sse/model/FindDiffUserDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class FindDiffUserDto {
-    private Integer diffIuser;
+    private Long diffIuser;
     private Integer reviewCount;
 }
