+++ src/main/java/com/team5/projrental/payment/model/proc/GetPaymentListDto.java
@@ -9,21 +9,21 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GetPaymentListDto {
-    private Integer iuser;
+    private Long iuser;
     private Integer role;
     private int flag;
     private int ipayment;
     private int page;
     private int paymentPerPage;
 
-    public GetPaymentListDto(int iuser, Integer role, int page, boolean hasRole) {
+    public GetPaymentListDto(Long iuser, Integer role, int page, boolean hasRole) {
         this.iuser = iuser;
         this.role = role;
         this.page = page;
         this.paymentPerPage = Const.PAY_PER_PAGE;
     }
 
-    public GetPaymentListDto(Integer iuser, int flag, int ipayment) {
+    public GetPaymentListDto(Long iuser, int flag, int ipayment) {
         this.iuser = iuser;
         this.flag = flag;
         this.ipayment = ipayment;
