+++ src/main/java/com/team5/projrental/payment/review/model/RiviewVo.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class RiviewVo {
     private int ipayment;
-    private Integer iuser;
+    private Long iuser;
 }
