+++ src/main/java/com/team5/projrental/product/ProductMapper.java
@@ -36,7 +36,7 @@ public interface ProductMapper {
     //
 
     //
-    int checkIuser(Integer iuser);
+    Long checkIuser(Long iuser);
     //
